backend/ai_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy==2.1.3
Pillow==11.0.0

